# Improve performance of include? by 5-10x