--- a/lib/ipaddr.rb
+++ b/lib/ipaddr.rb
@@ -176,9 +176,7 @@ class IPAddr
   def include?(other)
     other = coerce_other(other)
     return false unless other.family == family
-    range = to_range
-    other = other.to_range
-    range.begin <= other.begin && range.end >= other.end
+    begin_addr <= other.begin_addr && end_addr >= other.end_addr
   end
   alias === include?
 
@@ -400,17 +398,6 @@ class IPAddr
 
   # Creates a Range object for the network address.
   def to_range
-    begin_addr = (@addr & @mask_addr)
-
-    case @family
-    when Socket::AF_INET
-      end_addr = (@addr | (IN4MASK ^ @mask_addr))
-    when Socket::AF_INET6
-      end_addr = (@addr | (IN6MASK ^ @mask_addr))
-    else
-      raise AddressFamilyError, "unsupported address family"
-    end
-
     self.class.new(begin_addr, @family)..self.class.new(end_addr, @family)
   end
 
@@ -490,6 +477,21 @@ class IPAddr
   end
 
   protected
+
+  def begin_addr
+    @addr & @mask_addr
+  end
+
+  def end_addr
+    case @family
+    when Socket::AF_INET
+      @addr | (IN4MASK ^ @mask_addr)
+    when Socket::AF_INET6
+      @addr | (IN6MASK ^ @mask_addr)
+    else
+      raise AddressFamilyError, "unsupported address family"
+    end
+  end
 
   # Set +@addr+, the internal stored ip address, to given +addr+. The
   # parameter +addr+ is validated using the first +family+ member,