@skipkayhil 2022-12-16
I opened #48 to address the failing truffleruby test