@mitchellhenke 2023-01-18
I've seen similar in my own profiling, it would be great to see something like this merged!